Jumping is now more forgiving, making jumping on downward slopes reasonable.
This introduces `entity.CanJump` a separate boolean from `entity.Grounded`. The collision sensor for CanJump Extends 11 units further than a entity's feet. essentially, entities can jump if they are 11 units off the ground. Because of this, it is possible with a well timed input to exploit this and jump higher than intended. Jump power may have to be weakened, or a more sophisticated fix for jumping on downward slopes should be found.